+++ src/Controller/ItemsController.spec.ts
@@ -293,80 +293,4 @@ describe('ItemsController', () => {
     expect(result.statusCode).toEqual(200)
     expect(itemRepository.remove).toHaveBeenCalledWith(extension)
   })
-
-  it('should not delete email backups items if none are found', async () => {
-    request.body = {}
-    request.params = {
-      userUuid: '1-2-3',
-    }
-
-    itemRepository.findAll = jest.fn().mockReturnValue([])
-    itemRepository.remove = jest.fn()
-
-    const httpResponse = <results.OkResult> await createController().disableEmailBackups(request)
-    const result = await httpResponse.executeAsync()
-
-    expect(result.statusCode).toEqual(404)
-    expect(itemRepository.remove).not.toHaveBeenCalled()
-  })
-
-  it('should skip email backups items that do not have a content', async () => {
-    request.body = {}
-    request.params = {
-      userUuid: '1-2-3',
-    }
-
-    const extension = {
-      uuid: 'e-1-2-3',
-    } as jest.Mocked<Item>
-    itemRepository.findAll = jest.fn().mockReturnValue([extension])
-    itemRepository.remove = jest.fn()
-
-    const httpResponse = <results.OkResult> await createController().disableEmailBackups(request)
-    const result = await httpResponse.executeAsync()
-
-    expect(result.statusCode).toEqual(404)
-    expect(itemRepository.remove).not.toHaveBeenCalled()
-  })
-
-  it('should skip email backups items that do not have a subtype of backup.email_archive', async () => {
-    request.body = {}
-    request.params = {
-      userUuid: '1-2-3',
-    }
-
-    const extension = {
-      uuid: 'e-1-2-3',
-      content: 'something',
-    } as jest.Mocked<Item>
-    itemRepository.findAll = jest.fn().mockReturnValue([extension])
-    itemRepository.remove = jest.fn()
-
-    const httpResponse = <results.OkResult> await createController().disableEmailBackups(request)
-    const result = await httpResponse.executeAsync()
-
-    expect(result.statusCode).toEqual(404)
-    expect(itemRepository.remove).not.toHaveBeenCalled()
-  })
-
-  it('should disable email backups by deleting SF|Extension items', async () => {
-    request.body = {}
-    request.params = {
-      userUuid: '1-2-3',
-    }
-
-    const extension = {
-      uuid: 'e-1-2-3',
-      content: 'something',
-    } as jest.Mocked<Item>
-    itemRepository.findAll = jest.fn().mockReturnValue([extension])
-    itemRepository.remove = jest.fn()
-    contentDecoder.decode = jest.fn().mockReturnValue({ subtype: 'backup.email_archive' })
-
-    const httpResponse = <results.OkResult> await createController().disableEmailBackups(request)
-    const result = await httpResponse.executeAsync()
-
-    expect(result.statusCode).toEqual(200)
-    expect(itemRepository.remove).toHaveBeenCalledWith(extension)
-  })
 })
